# gen/http/openapi.yaml
swagger: "2.0"
info:
    title: starterpack
    description: Best API REST building with GoaDesign
    contact:
        name: Guillaume MORIN
        email: guillaume.morin@epitech.eu
        url: https://guillaume-morin.fr/
    license:
        name: GOA X GM
        url: https://guillaume-morin.fr/
    version: "1.0"
host: localhost:8088
consumes:
    - application/json
    - application/xml
    - application/gob
produces:
    - application/json
    - application/xml
    - application/gob
paths:
...
    /refresh:
        post:
            tags:
                - jwtToken
            summary: refresh jwtToken
            description: Refresh Token
            operationId: jwtToken#refresh
            parameters:
                - name: Authorization
                  in: header
                  description: OAuth token
                  required: false
                  type: string
                  pattern: ^Bearer [^ ]+$
                - name: RefreshRequestBody
                  in: body
                  required: true
                  schema:
                    $ref: '#/definitions/JWTTokenRefreshRequestBody'
                    required:
                        - refresh_token
            responses:
                "200":
                    description: OK response.
                    schema:
                        $ref: '#/definitions/JWTTokenRefreshResponseBody'
                        required:
                            - access_token
                            - refresh_token
                            - success
                "500":
                    description: Internal Server Error response.
                    schema:
                        $ref: '#/definitions/JWTTokenRefreshUnknownErrorResponseBody'
                        required:
                            - err
                            - success
                            - error_code
            schemes:
                - http
            security:
                - OAuth2_header_Authorization: []
    /signin:
        post:
            tags:
                - jwtToken
            summary: signin jwtToken
            description: signin
            operationId: jwtToken#signin
            parameters:
                - name: Authorization
                  in: header
                  description: OAuth token
                  required: false
                  type: string
                  pattern: ^Bearer [^ ]+$
                - name: SigninRequestBody
                  in: body
                  required: true
                  schema:
                    $ref: '#/definitions/JWTTokenSigninRequestBody'
                    required:
                        - password
                        - email
            responses:
                "200":
                    description: OK response.
                    schema:
                        $ref: '#/definitions/JWTTokenSigninResponseBody'
                        required:
                            - access_token
                            - refresh_token
                            - success
                "500":
                    description: Internal Server Error response.
                    schema:
                        $ref: '#/definitions/JWTTokenSigninUnknownErrorResponseBody'
                        required:
                            - err
                            - success
                            - error_code
            schemes:
                - http
            security:
                - OAuth2_header_Authorization: []
    /signup:
        post:
            tags:
                - jwtToken
            summary: signup jwtToken
            description: signup to generate jwt token
            operationId: jwtToken#signup
            parameters:
                - name: Authorization
                  in: header
                  description: OAuth token
                  required: false
                  type: string
                  pattern: ^Bearer [^ ]+$
                - name: SignupRequestBody
                  in: body
                  required: true
                  schema:
                    $ref: '#/definitions/JWTTokenSignupRequestBody'
                    required:
                        - email
                        - password
                        - confirm_password
            responses:
                "200":
                    description: OK response.
                    schema:
                        $ref: '#/definitions/JWTTokenSignupResponseBody'
                        required:
                            - access_token
                            - refresh_token
                            - success
                "500":
                    description: Internal Server Error response.
                    schema:
                        $ref: '#/definitions/JWTTokenSignupUnknownErrorResponseBody'
                        required:
                            - err
                            - success
                            - error_code
            schemes:
                - http
            security:
                - OAuth2_header_Authorization: []
...
definitions:
...
    JWTTokenRefreshRequestBody:
        title: JWTTokenRefreshRequestBody
        type: object
        properties:
            refresh_token:
                type: string
                example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiYWRtaW4iOnRydWV9.TJVA95OrM7E2cBab30RMHrHDcEfxjoYZgeFONFh7HgQ
                minLength: 32
        example:
            refresh_token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiYWRtaW4iOnRydWV9.TJVA95OrM7E2cBab30RMHrHDcEfxjoYZgeFONFh7HgQ
        required:
            - refresh_token
    JWTTokenRefreshResponseBody:
        title: JWTTokenRefreshResponseBody
        type: object
        properties:
            access_token:
                type: string
                example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiYWRtaW4iOnRydWV9.TJVA95OrM7E2cBab30RMHrHDcEfxjoYZgeFONFh7HgQ
            refresh_token:
                type: string
                example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiYWRtaW4iOnRydWV9.TJVA95OrM7E2cBab30RMHrHDcEfxjoYZgeFONFh7HgQ
            success:
                type: boolean
                example: true
        example:
            access_token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiYWRtaW4iOnRydWV9.TJVA95OrM7E2cBab30RMHrHDcEfxjoYZgeFONFh7HgQ
            refresh_token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiYWRtaW4iOnRydWV9.TJVA95OrM7E2cBab30RMHrHDcEfxjoYZgeFONFh7HgQ
            success: false
        required:
            - access_token
            - refresh_token
            - success
    JWTTokenRefreshUnknownErrorResponseBody:
        title: JWTTokenRefreshUnknownErrorResponseBody
        type: object
        properties:
            err:
                type: string
                example: sql no rows affected
            error_code:
                type: string
                example: TX_UPDATE_ITEM
            success:
                type: boolean
                default: false
                example: false
        description: Error not identified (500)
        example:
            err: sql no rows affected
            error_code: TX_UPDATE_ITEM
            success: true
        required:
            - err
            - success
            - error_code
    JWTTokenSigninRequestBody:
        title: JWTTokenSigninRequestBody
        type: object
        properties:
            email:
                type: string
                example: guillaume@epitech.eu
                format: email
            password:
                type: string
                description: Minimum 9 charactères / Chiffre Obligatoire
                example: JeSuisUnTest974
                pattern: \d
                minLength: 9
        example:
            email: guillaume@epitech.eu
            password: JeSuisUnTest974
        required:
            - password
            - email
    JWTTokenSigninResponseBody:
        title: JWTTokenSigninResponseBody
        type: object
        properties:
            access_token:
                type: string
                example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiYWRtaW4iOnRydWV9.TJVA95OrM7E2cBab30RMHrHDcEfxjoYZgeFONFh7HgQ
            refresh_token:
                type: string
                example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiYWRtaW4iOnRydWV9.TJVA95OrM7E2cBab30RMHrHDcEfxjoYZgeFONFh7HgQ
            success:
                type: boolean
                example: false
        example:
            access_token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiYWRtaW4iOnRydWV9.TJVA95OrM7E2cBab30RMHrHDcEfxjoYZgeFONFh7HgQ
            refresh_token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiYWRtaW4iOnRydWV9.TJVA95OrM7E2cBab30RMHrHDcEfxjoYZgeFONFh7HgQ
            success: false
        required:
            - access_token
            - refresh_token
            - success
    JWTTokenSigninUnknownErrorResponseBody:
        title: JWTTokenSigninUnknownErrorResponseBody
        type: object
        properties:
            err:
                type: string
                example: sql no rows affected
            error_code:
                type: string
                example: TX_UPDATE_ITEM
            success:
                type: boolean
                default: false
                example: false
        description: Error not identified (500)
        example:
            err: sql no rows affected
            error_code: TX_UPDATE_ITEM
            success: false
        required:
            - err
            - success
            - error_code
    JWTTokenSignupRequestBody:
        title: JWTTokenSignupRequestBody
        type: object
        properties:
            confirm_password:
                type: string
                description: Minimum 9 charactères / Chiffre Obligatoire
                example: JeSuisUnTest974
                pattern: \d
                minLength: 9
            email:
                type: string
                example: guillaume@epitech.eu
                format: email
            password:
                type: string
                description: Minimum 9 charactères / Chiffre Obligatoire
                example: JeSuisUnTest974
                pattern: \d
                minLength: 9
        example:
            confirm_password: JeSuisUnTest974
            email: guillaume@epitech.eu
            password: JeSuisUnTest974
        required:
            - email
            - password
            - confirm_password
    JWTTokenSignupResponseBody:
        title: JWTTokenSignupResponseBody
        type: object
        properties:
            access_token:
                type: string
                example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiYWRtaW4iOnRydWV9.TJVA95OrM7E2cBab30RMHrHDcEfxjoYZgeFONFh7HgQ
            refresh_token:
                type: string
                example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiYWRtaW4iOnRydWV9.TJVA95OrM7E2cBab30RMHrHDcEfxjoYZgeFONFh7HgQ
            success:
                type: boolean
                example: true
        example:
            access_token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiYWRtaW4iOnRydWV9.TJVA95OrM7E2cBab30RMHrHDcEfxjoYZgeFONFh7HgQ
            refresh_token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiYWRtaW4iOnRydWV9.TJVA95OrM7E2cBab30RMHrHDcEfxjoYZgeFONFh7HgQ
            success: true
        required:
            - access_token
            - refresh_token
            - success
    JWTTokenSignupUnknownErrorResponseBody:
        title: JWTTokenSignupUnknownErrorResponseBody
        type: object
        properties:
            err:
                type: string
                example: sql no rows affected
            error_code:
                type: string
                example: TX_UPDATE_ITEM
            success:
                type: boolean
                default: false
                example: true
        description: Error not identified (500)
        example:
            err: sql no rows affected
            error_code: TX_UPDATE_ITEM
            success: false
        required:
            - err
            - success
            - error_code
...
securityDefinitions:
    OAuth2_header_Authorization:
        type: oauth2
        description: Use OAuth2 to authenticate
        flow: application
        tokenUrl: /authorization
        scopes:
            api:read: Provides read access
    jwt_header_jwtToken:
        type: apiKey
        description: |-
            Secures endpoint by requiring a valid JWT token retrieved via the signin endpoint. Supports scopes "api:read" and "api:write".

            **Security Scopes**:
              * `api:read`: Read-only access
              * `api:write`: Read and write access
        name: jwtToken
        in: header